server: extract static + manifest, readBody to route-helpers
- src/server/static.ts: MIME_TYPES, loadManifest, serveStatic
- src/server/routes/route-helpers.ts (new skeleton): readBody, json, jsonError

Net: server.ts -85 lines (7971 → 7886). No behaviour change.
